Move the log collector and intrusion detection controllers into pkg/enterprise
The Typha and node TLS helpers move to the shared controller utilities, which
is what lets the intrusion detection controller stop importing the installation
controller.
