+++ pkg/enterprise/controller/intrusiondetection/intrusiondetection_controller_test.go
@@ -49,10 +49,10 @@ import (
 	"github.com/tigera/operator/pkg/controller/utils"
 	ctrlrfake "github.com/tigera/operator/pkg/ctrlruntime/client/fake"
 	entcertificatemanager "github.com/tigera/operator/pkg/enterprise/certificatemanager"
+	"github.com/tigera/operator/pkg/enterprise/render/intrusiondetection/dpi"
 	"github.com/tigera/operator/pkg/render"
 	relasticsearch "github.com/tigera/operator/pkg/render/common/elasticsearch"
 	rtest "github.com/tigera/operator/pkg/render/common/test"
-	"github.com/tigera/operator/pkg/render/intrusiondetection/dpi"
 	"github.com/tigera/operator/pkg/render/logstorage/eck"
 	"github.com/tigera/operator/test"
 )
