+++ pkg/enterprise/controller/intrusiondetection/intrusiondetection_suite_test.go
@@ -1,10 +1,10 @@
-// Copyright (c) 2026 Tigera, Inc. All rights reserved.
+// Copyright (c) 2020-2026 Tigera, Inc. All rights reserved.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//	http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package otelcollector_test
+package intrusiondetection
 
 import (
 	"testing"
@@ -21,9 +21,9 @@ import (
 	"github.com/onsi/gomega"
 )
 
-func TestRender(t *testing.T) {
+func TestStatus(t *testing.T) {
 	gomega.RegisterFailHandler(ginkgo.Fail)
 	suiteConfig, reporterConfig := ginkgo.GinkgoConfiguration()
-	reporterConfig.JUnitReport = "../../../report/ut/otelcollector_render_suite.xml"
-	ginkgo.RunSpecs(t, "pkg/render/otelcollector Suite", suiteConfig, reporterConfig)
+	reporterConfig.JUnitReport = "../../../../report/ut/intrusiondetection_controller_suite.xml"
+	ginkgo.RunSpecs(t, "pkg/enterprise/controller/intrusiondetection Controller Suite", suiteConfig, reporterConfig)
 }
