+++ pkg/enterprise/render/otelcollector/collector-config.yaml.template
