+++ pkg/enterprise/render/otelcollector/component.go
